Im entering the scene :dancer:
Yo!  Kk, so the newGameview we are using right now assumes every time we
call it, we're starting a new round of moves.  I don't think this is
meant to happen (say its Dracula's turn and we call gameview, things
will be set to godlamings turn.  I just left a comment in there but I'll
get working on it today.