Made whoseTurn Function, added stdio
Added a function to get the ID of the last player in the past plays
string.  Also...we didn't have stdio XD erm, I think we need it.
Anyways, things seemed to compile fine on putty but I couldn't smoothly
copy all the putty changes into my git on windows so there might be a
few small issues.