Added getTurnNo
I added a function which gets the current turn number.
The past plays string omits the trailing space on the last play in the
string.  As such, I just add 1 to the size of the string (to make up for
it) and divide the result by 8 to get the result.